584 user story digital avatar increase size limit for rag documents (#586)
